requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-ckeditor==6.4.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-phone-field==1.8.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
